+++ src/boxrec-pages/event/bout/boxrec.page.event.bout.spec.ts
@@ -1,5 +1,5 @@
 import {mockBoutCaneloGGG1} from "boxrec-mocks";
-import {BoxrecBasic, WinLossDraw} from "../../boxrec.constants";
+import {WinLossDraw} from "../../boxrec.constants";
 import {WeightDivision} from "../../champions/boxrec.champions.constants";
 import {BoxingBoutOutcome} from "../boxrec.event.constants";
 import {BoutPageLast6, BoxrecEventBoutOutput} from "./boxrec.event.bout.constants";