+++ src/boxrec-pages/profile/boxrec.profile.constants.ts
@@ -54,7 +54,6 @@ export enum BoxrecProfileTable {
     birthPlace = "birth place",
     company = "company",
     registeredContact = "registered contact",
-    death = "death date",
 }
 
 export interface BoxrecProfileBoutLocation {